+++ tests/test_trans_ir_bridge_rc_load.rs
@@ -16,7 +16,7 @@ fn test_trans_ir_bridge_rc_1kHz() {
 }
 
 #[allow(dead_code)]
-fn build(amp: f32, freq: f32, isat: f32) -> Circuit {
+fn build(amp: f64, freq: f64, isat: f64) -> Circuit {
     let mut ckt = Circuit::new();
 
     // bridge input voltage
