+++ tests/test_trans_irrc.rs
@@ -86,7 +86,7 @@ fn test_irrc_trans_lpf_loop() {
 }
 
 
-fn build( freq: f32 ) -> Circuit {
+fn build( freq: f64 ) -> Circuit {
     let mut ckt = Circuit::new();
 
     // 10V Voltage Source
